refactoring towards more injection

+++ serverless_proto/postcard/src/main/java/com/serverless/Environment.java
@@ -8,7 +8,7 @@ public interface Environment {
 
   String getLobApiVersion();
 
-  Integer getPostcardPrice();
+  Long getPostcardPrice();
 
   String getPostcardCurrancy();
